마이페이지 수정 , 쪽지 구현 완료 , 비밀번호 변경 추가 구현 완료

+++ conbee/src/main/resources/static/js/note/noteKeep.js
@@ -1,5 +1,4 @@
 
-
 const messageName = document.getElementById("messageName");
 const messageContent = document.getElementById("messageContent");
 const messageDate = document.getElementById("messageDate");
@@ -46,6 +45,12 @@ document.getElementById('deleteBtn').addEventListener('click', e =>{
       console.log(messageNoList);
     }
   }
+
+  if(messageNoList.length==0){
+    alert("삭제할 쪽지를 선택해주세요");
+    e.preventDefault();
+    return;            
+  }
   if(confirm("정말 삭제??")){
   
     location.href="/note/deleteNoteKeep?messageNoList=" + messageNoList;
@@ -53,3 +58,27 @@ document.getElementById('deleteBtn').addEventListener('click', e =>{
   }
 
 });
+
+/* 검색 */
+const keepName = document.getElementById("keepName");
+const keepNameBtn = document.getElementById("keepNameBtn");
+
+const url = new URL(location.href);
+const urlParams = url.searchParams;
+
+keepNameBtn.addEventListener("click", ()=>{
+    const url = new URL(location.href);
+    const urlParams = url.searchParams;
+    location.href = "/note/note-keep?query=" + keepName.value;
+});
+
+
+/* 검색 기록 남기기 */
+(()=>{
+  const searchWord = urlParams.get("query"); // 검색어
+
+  // 검색을 했을 경우
+  if(searchWord !=null){
+    keepName.value = searchWord; // 검색어를 input에 추가
+  }
+})();
